Move gear-tracker into a new directory
